+++ lib/src/core/app/app.dart
@@ -1,6 +1,6 @@
 import 'package:bot_toast/bot_toast.dart';
 import 'package:enigma/src/core/router/router.dart';
-import 'package:enigma/src/core/styles/theme/enigma_theme.dart';
+import 'package:enigma/src/core/styles/theme/app_theme.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
@@ -17,9 +17,9 @@ class App extends ConsumerWidget {
       routerDelegate: router.routerDelegate,
       routeInformationParser: router.routeInformationParser,
       routeInformationProvider: router.routeInformationProvider,
-      darkTheme: customDarkTheme,
-      theme: customLightTheme,
-      themeMode: ThemeMode.light,
+      darkTheme: darkTheme,
+      theme: lightTheme,
+      themeMode: ThemeMode.dark,
     );
   }
 }
