+++ lib/src/features/auth/presentation/auth_screen/view/auth_screen.dart
@@ -17,7 +17,8 @@ class AuthScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: Theme.of(context).primaryColor,
+      resizeToAvoidBottomInset: false,
+      backgroundColor: Theme.of(context).colorScheme.secondary,
       appBar: AppBar(
         backgroundColor: Colors.transparent,
         title: Row(
@@ -25,16 +26,13 @@ class AuthScreen extends StatelessWidget {
           children: [
             Text(
               "E ",
-              style: Theme.of(context)
-                  .textTheme
-                  .displaySmall
-                  ?.copyWith(color: Colors.white),
+              style: Theme.of(context).textTheme.displaySmall,
             ),
             Text(
               "Enigma",
               style: TextStyle(
-                color: Theme.of(context).secondaryHeaderColor,
-              ),
+                  //color: Theme.of(context).secondaryHeaderColor,
+                  ),
             ),
           ],
         ),
@@ -46,19 +44,13 @@ class AuthScreen extends StatelessWidget {
           children: [
             Text(
               'Connect friends easily & quickly',
-              style: Theme.of(context)
-                  .textTheme
-                  .displayLarge
-                  ?.copyWith(color: Colors.white),
+              style: Theme.of(context).textTheme.displayLarge,
             ),
             Padding(
               padding: const EdgeInsets.all(8.0),
               child: Text(
                 'Our chat app is the perfect way to stay connected with friends and family.',
-                style: Theme.of(context)
-                    .textTheme
-                    .titleLarge
-                    ?.copyWith(color: Colors.white),
+                style: Theme.of(context).textTheme.titleLarge,
               ),
             ),
             Padding(
@@ -82,8 +74,8 @@ class AuthScreen extends StatelessWidget {
               ),
             ),
             OrWidget(
-              color: Theme.of(context).splashColor,
-            ),
+                //color: Theme.of(context).splashColor,
+                ),
             Padding(
               padding: const EdgeInsets.all(8.0),
               child: CustomElevatedButton(
@@ -98,10 +90,10 @@ class AuthScreen extends StatelessWidget {
                 context.push(LoginScreen.setRoute());
               },
               child: RichText(
-                text: const TextSpan(children: [
+                text: TextSpan(children: [
                   TextSpan(text: "Existing account?"),
                   TextSpan(text: "  Log in")
-                ]),
+                ], style: Theme.of(context).textTheme.bodySmall),
               ),
             )
           ],
