+++ lib/src/features/auth/presentation/components/custom_elevated_button.dart
@@ -10,16 +10,18 @@ class CustomElevatedButton extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return ElevatedButton(
-      onPressed: onPressed,
-      style: ElevatedButton.styleFrom(
-        backgroundColor: Colors.white,
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(15)),
-        minimumSize: Size(context.width * 0.8, context.width * 0.15),
-      ),
-      child: Text(
-        buttonName,
-        style: Theme.of(context).textTheme.titleLarge,
+    return SizedBox(
+      width: context.width * 0.8,
+      height: context.height * 0.07,
+      child: ElevatedButton(
+        onPressed: onPressed,
+        // style: ElevatedButton.styleFrom(
+        //   minimumSize: Size(context.width * 0.8, context.width * 0.15),
+        // ),
+        child: Text(
+          buttonName,
+          //style: Theme.of(context).textTheme.titleLarge,
+        ),
       ),
     );
   }
